+++ src/tasks/Assignments.test.ts
@@ -1,4 +1,5 @@
 import tape from '../test/tape-seneca'
+import {spy} from 'sinon'
 import Assignments from './Assignments'
 
 function mockShifts() {
@@ -21,7 +22,7 @@ test('create', async function(t) {
     shiftKey: 'shiftOne',
   }
   const response = await this.act('role:Assignments,cmd:create', {values})
-  t.ok(response.key)
+  t.ok(response.key, 'it returns the key')
 })
 
 test('update', async function(t) {
@@ -36,5 +37,23 @@ test('update', async function(t) {
   t.ok(response.key)
 
   const {assignment} = await this.act('role:Firebase,cmd:get', {assignment: key})
-  t.ok(assignment.newValue)
+  t.ok(assignment.newValue, 'it sets the new value')
+})
+
+test('shift changes', async function(t) {
+  const uid = 'abc123'
+  const shiftChangeSpy = spy()
+  this.add('role:ShiftChanges,cmd:create', shiftChangeSpy)
+
+  const {key} = await this.act('role:Assignments,cmd:create', {uid, values: {
+    oppKey: 'testOpp',
+    engagementKey: 'testEngagement',
+    teamKey: 'testTeam',
+    shiftKey: 'shiftOne',
+  }})
+
+  await this.act('role:Assignments,cmd:update', {uid, key, values: {shiftKey: 'shiftTwo'}})
+  await this.act('role:Assignments,cmd:remove', {uid, key})
+
+  t.equals(shiftChangeSpy.callCount, 3)
 })