+++ src/tasks/Shifts.test.ts
@@ -5,23 +5,33 @@ import Shifts from './Shifts'
 const test = tape('Shifts', [Shifts])
 
 test('remove', async function(t) {
-  let shift, assignments
+  const uid = 'abc123'
 
   // Check env
-  ({shift, assignments} =
-    await this.act('role:Firebase,cmd:get', {shift: 'shiftOne', assignments: {shiftKey: 'shiftOne'}}))
-  t.ok(shift)
-  t.equals(assignments.length, 1)
+  const {shift: originalShift, assignments: originalAssignments} =
+    await this.act('role:Firebase,cmd:get', {shift: 'shiftOne', assignments: {shiftKey: 'shiftOne'}})
+  t.ok(originalShift)
+  t.equals(originalAssignments.length, 1)
+
+  const createShiftChangeSpy = spy()
+  this.add('role:ShiftChanges,cmd:create', createShiftChangeSpy)
 
   // Action
-  await this.act('role:Shifts,cmd:remove,key:shiftOne');
+  await this.act('role:Shifts,cmd:remove,key:shiftOne', {uid})
 
   // Spec
-  ({shift, assignments} =
-    await this.act('role:Firebase,cmd:get', {shift: 'shiftOne', assignments: {shiftKey: 'shiftOne'}}))
+  const {shift, assignments} =
+    await this.act('role:Firebase,cmd:get', {shift: 'shiftOne', assignments: {shiftKey: 'shiftOne'}})
 
   t.notOk(shift, 'Shift should be deleted')
   t.equals(assignments.length, 0, 'Shift\'s assignments should be deleted')
+
+  t.equals(createShiftChangeSpy.callCount, 1, 'shift change is created for the assignment')
+  t.ok(createShiftChangeSpy.calledWithMatch(match({
+    shift: originalShift,
+    assignment: originalAssignments[0],
+    uid,
+  })), 'shift change is given the shift, assignment and uid')
 })
 
 test('create', async function(t) {